Move FunctionIO to apis/apiextensions/fn/io/
